Languages can parse source text themselves
A language can now specify, using a package variable, that it wants the
source text.  An axk script can contain multiple Ln declarations, and
the text after an Ln declaration and before the next can be given to
that package as a single-quoted heredoc.  See example in ex/l0.axk.